# apps/api/openapi/user.openapi.json
{
  "components": {
    "schemas": {
...
      "RecordReadingActivityDto": {
        "properties": {
          "documentId": {
            "type": "string"
          },
          "eventKind": {
            "type": "string"
          },
          "offset": {
            "minimum": 0,
            "type": "number"
          },
          "progress": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "revisionId": {
            "type": "string"
          }
        },
        "required": ["documentId", "eventKind"],
        "type": "object"
      },
...
      "ResolveSelectionDto": {
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": ["text"],
        "type": "object"
      },
      "SaveReadingDto": {
        "properties": {
          "documentId": {
            "type": "string"
          },
          "releaseId": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          },
          "targetKind": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sylis_session": {
        "in": "cookie",
        "name": "sylis_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Sylis user and administration HTTP API",
    "title": "Sylis User API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/lexicon/entries/{id}": {
      "get": {
        "operationId": "LexiconController_entry",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Lexicon"]
      }
    },
    "/api/v1/lexicon/entries/{id}/materials": {
      "get": {
        "operationId": "LexiconController_entryMaterials",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Lexicon"]
      }
    },
    "/api/v1/lexicon/headwords/{id}": {
      "get": {
        "operationId": "LexiconController_headword",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Lexicon"]
      }
    },
...
    "/api/v1/lexicon/senses/{id}": {
      "get": {
        "operationId": "LexiconController_sense",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Lexicon"]
      }
    },
    "/api/v1/lexicon/senses/{id}/materials": {
      "get": {
        "operationId": "LexiconController_senseMaterials",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Lexicon"]
      }
    },
...
    "/api/v1/reading/activities": {
      "post": {
        "operationId": "ReadingController_activity",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordReadingActivityDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Reading"]
      }
    },
    "/api/v1/reading/documents/{id}": {
      "get": {
        "operationId": "ReadingController_document",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Reading"]
      }
    },
...
    "/api/v1/reading/revisions/{id}/annotations": {
      "get": {
        "operationId": "ReadingController_annotations",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Reading"]
      }
    },
    "/api/v1/reading/revisions/{id}/resolve-selection": {
      "post": {
        "operationId": "ReadingController_resolve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveSelectionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Reading"]
      }
    },
    "/api/v1/reading/saved": {
      "get": {
        "operationId": "ReadingController_saved",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Reading"]
      },
      "post": {
        "operationId": "ReadingController_save",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveReadingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Reading"]
      }
    },
    "/api/v1/reading/saved/{id}": {
      "delete": {
        "operationId": "ReadingController_unsave",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["Reading"]
      }
    },
...
  },
  "servers": [],
...
}